+++ analysis/import/CSVImporter/src/test/kotlin/de/maibornwolff/codecharta/importer/csv/ParserDialogTest.kt
@@ -26,14 +26,15 @@ class ParserDialogTest {
     fun `should output correct arguments`() {
         val fileName = "in.csv"
         val outputFileName = "out.cc.json"
+        val pathColumnName = "path"
         val delimiter = ";"
         val pathSeparator = "/"
         val isCompressed = false
 
         mockkStatic("com.github.kinquirer.components.InputKt")
         every {
-            KInquirer.promptInput(any(), any(), any(), any())
-        } returns fileName andThen "" andThen outputFileName andThen delimiter andThen pathSeparator
+            KInquirer.promptInput(any(), any(), any(), any(), any())
+        } returns fileName andThen "" andThen outputFileName andThen pathColumnName andThen delimiter andThen pathSeparator
         mockkStatic("com.github.kinquirer.components.ConfirmKt")
         every {
             KInquirer.promptConfirm(any(), any())
@@ -43,7 +44,9 @@ class ParserDialogTest {
 
         val cmdLine = CommandLine(CSVImporter())
         val parseResult = cmdLine.parseArgs(*parserArguments.toTypedArray())
-        Assertions.assertThat(parseResult.matchedOption("output-file").getValue<String>().equals(outputFileName))
+        Assertions.assertThat(parseResult.matchedOption("output-file").getValue<String>()
+                .equals(outputFileName))
+        Assertions.assertThat(parseResult.matchedOption("path-column-name").getValue<String>()).isEqualTo(pathColumnName)
         Assertions.assertThat(parseResult.matchedOption("delimiter").getValue<Char>()).isEqualTo(delimiter[0])
         Assertions.assertThat(parseResult.matchedOption("path-separator").getValue<Char>()).isEqualTo(pathSeparator[0])
         Assertions.assertThat(parseResult.matchedOption("not-compressed").getValue<Boolean>()).isEqualTo(isCompressed)
@@ -56,6 +59,7 @@ class ParserDialogTest {
         val fileName2 = "in2.csv"
         val fileName3 = "in3.csv"
         val outputFileName = "out.cc.json"
+        val pathColumnName = "path"
         val delimiter = ";"
         val pathSeparator = "/"
         val isCompressed = true
@@ -63,7 +67,7 @@ class ParserDialogTest {
         mockkStatic("com.github.kinquirer.components.InputKt")
         every {
             KInquirer.promptInput(any(), any(), any(), any(), any(), any())
-        } returns fileName andThen fileName2 andThen fileName3 andThen "" andThen outputFileName andThen delimiter andThen pathSeparator
+        } returns fileName andThen fileName2 andThen fileName3 andThen "" andThen outputFileName andThen pathColumnName andThen delimiter andThen pathSeparator
         mockkStatic("com.github.kinquirer.components.ConfirmKt")
         every {
             KInquirer.promptConfirm(any(), any())
@@ -74,6 +78,7 @@ class ParserDialogTest {
         val cmdLine = CommandLine(CSVImporter())
         val parseResult = cmdLine.parseArgs(*parserArguments.toTypedArray())
         Assertions.assertThat(parseResult.matchedOption("output-file").getValue<String>().equals(outputFileName))
+        Assertions.assertThat(parseResult.matchedOption("path-column-name").getValue<String>()).isEqualTo(pathColumnName)
         Assertions.assertThat(parseResult.matchedOption("delimiter").getValue<Char>()).isEqualTo(delimiter[0])
         Assertions.assertThat(parseResult.matchedOption("path-separator").getValue<Char>()).isEqualTo(pathSeparator[0])
         assertNull(parseResult.matchedOption("not-compressed"))